MB-24291 Delete pidfile on graceful shutdown
Due to the requirements of the System V init scripts on Centos 6 and other
allied platforms  we write the pid of the couchbbase server to the
couchbase-server.pid file. This file is not deleted on graceful shutdown on
the couchbase server.
In order to fix this we spawn a new process during init of the ns_server_sup
which is tasked only with waiting for an exit signal when the parent is in
the process of terminating and deleting the pidfile when it receives
the signal. We do this only on the affected platforms.

Change-Id: I5aa58e7aa21600ef48af7b7fd34eeaaf4eda44e1

Signed-off-by: Amit Krishnan <krish.amit@gmail.com>